Remove filterFunctions from PageSourceOptions
