+++ presto-hive/src/main/java/com/facebook/presto/hive/HivePageSource.java
@@ -15,11 +15,9 @@ package com.facebook.presto.hive;
 
 import com.facebook.presto.hive.HivePageSourceProvider.BucketAdaptation;
 import com.facebook.presto.hive.HivePageSourceProvider.ColumnMapping;
-import com.facebook.presto.hive.HivePageSourceProvider.ColumnMappingKind;
 import com.facebook.presto.spi.ConnectorPageSource;
 import com.facebook.presto.spi.Page;
 import com.facebook.presto.spi.PageSourceOptions;
-import com.facebook.presto.spi.PageSourceOptions.AbstractFilterFunction;
 import com.facebook.presto.spi.PrestoException;
 import com.facebook.presto.spi.block.ArrayBlock;
 import com.facebook.presto.spi.block.Block;
@@ -38,7 +36,6 @@ import com.facebook.presto.spi.type.Type;
 import com.facebook.presto.spi.type.TypeManager;
 import com.facebook.presto.spi.type.VarcharType;
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.collect.ImmutableList;
 import it.unimi.dsi.fastutil.ints.IntArrayList;
 import org.apache.hadoop.hive.serde2.typeinfo.ListTypeInfo;
 import org.apache.hadoop.hive.serde2.typeinfo.MapTypeInfo;
@@ -51,7 +48,6 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
-import java.util.Set;
 import java.util.function.Function;
 
 import static com.facebook.presto.hive.HiveBucketing.getHiveBucket;
@@ -88,7 +84,6 @@ import static com.facebook.presto.spi.StandardErrorCode.NOT_SUPPORTED;
 import static com.facebook.presto.spi.block.ColumnarArray.toColumnarArray;
 import static com.facebook.presto.spi.block.ColumnarMap.toColumnarMap;
 import static com.facebook.presto.spi.block.ColumnarRow.toColumnarRow;
-import static com.facebook.presto.spi.predicate.Utils.nativeValueToBlock;
 import static com.facebook.presto.spi.type.BigintType.BIGINT;
 import static com.facebook.presto.spi.type.BooleanType.BOOLEAN;
 import static com.facebook.presto.spi.type.Chars.isCharType;
@@ -103,7 +98,6 @@ import static com.facebook.presto.spi.type.TimestampType.TIMESTAMP;
 import static com.facebook.presto.spi.type.TinyintType.TINYINT;
 import static com.facebook.presto.spi.type.Varchars.isVarcharType;
 import static com.google.common.collect.ImmutableList.toImmutableList;
-import static com.google.common.collect.ImmutableSet.toImmutableSet;
 import static io.airlift.slice.Slices.utf8Slice;
 import static java.lang.Float.intBitsToFloat;
 import static java.lang.Math.max;
@@ -122,9 +116,7 @@ public class HivePageSource
 
     private final ConnectorPageSource delegate;
     private boolean filterAndProjectPushedDown;
-    private boolean filterOnPrefilledValuesFailed;
     private int[] outputChannels;
-    private Optional<Throwable> filterOnPrefilledValuesException = Optional.empty();
     private final boolean needsColumnMapping;
 
     public HivePageSource(
@@ -242,21 +234,10 @@ public class HivePageSource
     @Override
     public Page getNextPage()
     {
-        if (filterOnPrefilledValuesFailed) {
-            return null;
-        }
-
         try {
             Page dataPage = delegate.getNextPage();
             if (dataPage == null) {
                 return null;
-            }
-
-            if (dataPage.getPositionCount() > 0 && filterOnPrefilledValuesException.isPresent()) {
-                if (filterOnPrefilledValuesException.get() instanceof PrestoException) {
-                    throw (PrestoException) filterOnPrefilledValuesException.get();
-                }
-                throw new RuntimeException(filterOnPrefilledValuesException.get());
             }
 
             if (filterAndProjectPushedDown) {
@@ -786,48 +767,13 @@ public class HivePageSource
     @Override
     public boolean pushdownFilterAndProjection(PageSourceOptions options)
     {
-        List<AbstractFilterFunction> remainingFilterFunctions = pushdownFilterOnPrefilledColumns(options);
-
-        if (filterOnPrefilledValuesFailed) {
-            return true;
-        }
-
         this.outputChannels = Arrays.copyOf(options.getOutputChannels(), options.getOutputChannels().length);
 
-        filterAndProjectPushedDown = delegate.pushdownFilterAndProjection(createPageSourceOptionsForDelegate(options, remainingFilterFunctions));
+        filterAndProjectPushedDown = delegate.pushdownFilterAndProjection(createPageSourceOptionsForDelegate(options));
         return filterAndProjectPushedDown;
     }
 
-    private List<AbstractFilterFunction> pushdownFilterOnPrefilledColumns(PageSourceOptions options)
-    {
-        ImmutableList.Builder<AbstractFilterFunction> remainingFilterFunctions = ImmutableList.builder();
-        for (AbstractFilterFunction filterFunction : options.getFilterFunctions()) {
-            Set<ColumnMappingKind> inputKinds = getInputKinds(filterFunction);
-            if (!inputKinds.contains(PREFILLED)) {
-                remainingFilterFunctions.add(filterFunction);
-                continue;
-            }
-
-            if (inputKinds.size() > 1) {
-                // TODO Pass necessary pre-filled values to the delegate so it can evaluate these filters
-                throw new UnsupportedOperationException("Filters on a mix of partition and non-partition columns are not supported");
-            }
-
-            if (!filterFunction.isDeterministic()) {
-                throw new UnsupportedOperationException("Non-deterministic filters on partition columns are not supported");
-            }
-
-            if (!evaluateFilterFunctionOnPrefilledValues(filterFunction)) {
-                filterOnPrefilledValuesFailed = true;
-                close();
-                break;
-            }
-        }
-
-        return remainingFilterFunctions.build();
-    }
-
-    private PageSourceOptions createPageSourceOptionsForDelegate(PageSourceOptions options, List<AbstractFilterFunction> filterFunctions)
+    private PageSourceOptions createPageSourceOptionsForDelegate(PageSourceOptions options)
     {
         // Remove non-regular and non-interim columns from internal and output channels
         int[] internalChannels = Arrays.copyOf(options.getInternalChannels(), options.getInternalChannels().length);
@@ -846,32 +792,6 @@ public class HivePageSource
                 internalChannels,
                 outputChannels,
                 options.getReusePages(),
-                filterFunctions.toArray(new AbstractFilterFunction[0]),
                 options.getTargetBytes());
-    }
-
-    private Set<ColumnMappingKind> getInputKinds(AbstractFilterFunction filterFunction)
-    {
-        return Arrays.stream(filterFunction.getInputChannels())
-                .mapToObj(columnMappings::get)
-                .map(ColumnMapping::getKind)
-                .collect(toImmutableSet());
-    }
-
-    private boolean evaluateFilterFunctionOnPrefilledValues(AbstractFilterFunction filterFunction)
-    {
-        int[] inputChannels = filterFunction.getInputChannels();
-        Block[] blocks = new Block[inputChannels.length];
-        for (int i = 0; i < inputChannels.length; i++) {
-            blocks[i] = nativeValueToBlock(types[inputChannels[i]], prefilledValues[inputChannels[i]]);
-        }
-        PageSourceOptions.ErrorSet errorSet = new PageSourceOptions.ErrorSet();
-        int[] outputRows = new int[1];
-        int numRows = filterFunction.filter(new Page(blocks), outputRows, errorSet);
-        if (!errorSet.isEmpty()) {
-            filterOnPrefilledValuesException = Optional.of(errorSet.getFirstError(1));
-            return true;
-        }
-        return numRows == 1;
     }
 }
