+++ presto-main/src/main/java/com/facebook/presto/operator/TableScanOperator.java
@@ -19,7 +19,6 @@ import com.facebook.presto.spi.ColumnHandle;
 import com.facebook.presto.spi.ConnectorPageSource;
 import com.facebook.presto.spi.Page;
 import com.facebook.presto.spi.PageSourceOptions;
-import com.facebook.presto.spi.PageSourceOptions.AbstractFilterFunction;
 import com.facebook.presto.spi.UpdatablePageSource;
 import com.facebook.presto.spi.plan.PlanNodeId;
 import com.facebook.presto.split.EmptySplit;
@@ -37,9 +36,11 @@ import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Supplier;
 
+import static com.facebook.presto.SystemSessionProperties.getFilterAndProjectMinOutputPageSize;
 import static com.facebook.presto.SystemSessionProperties.isAriaScanEnabled;
 import static com.google.common.base.Preconditions.checkState;
 import static io.airlift.concurrent.MoreFutures.toListenableFuture;
+import static java.lang.Math.toIntExact;
 import static java.util.Objects.requireNonNull;
 
 public class TableScanOperator
@@ -279,8 +280,7 @@ public class TableScanOperator
                 channels,
                 channels,
                 reusePages,
-                new AbstractFilterFunction[0],
-                512 * 1024);
+                toIntExact(getFilterAndProjectMinOutputPageSize(operatorContext.getSession()).toBytes()));
         source.pushdownFilterAndProjection(options);
     }
