+++ presto-orc/src/main/java/com/facebook/presto/orc/reader/StructStreamReader.java
@@ -14,6 +14,7 @@
 package com.facebook.presto.orc.reader;
 
 import com.facebook.presto.memory.context.AggregatedMemoryContext;
+import com.facebook.presto.orc.AbstractFilterFunction;
 import com.facebook.presto.orc.ColumnGroupReader;
 import com.facebook.presto.orc.Filter;
 import com.facebook.presto.orc.Filters;
@@ -22,7 +23,6 @@ import com.facebook.presto.orc.StreamDescriptor;
 import com.facebook.presto.orc.metadata.ColumnEncoding;
 import com.facebook.presto.orc.stream.BooleanInputStream;
 import com.facebook.presto.orc.stream.InputStreamSources;
-import com.facebook.presto.spi.PageSourceOptions.AbstractFilterFunction;
 import com.facebook.presto.spi.Subfield;
 import com.facebook.presto.spi.Subfield.NestedField;
 import com.facebook.presto.spi.block.Block;
