+++ presto-orc/src/test/java/com/facebook/presto/orc/OrcTester.java
@@ -20,7 +20,6 @@ import com.facebook.presto.metadata.FunctionManager;
 import com.facebook.presto.orc.metadata.CompressionKind;
 import com.facebook.presto.spi.Page;
 import com.facebook.presto.spi.PageSourceOptions;
-import com.facebook.presto.spi.PageSourceOptions.AbstractFilterFunction;
 import com.facebook.presto.spi.block.Block;
 import com.facebook.presto.spi.block.BlockBuilder;
 import com.facebook.presto.spi.type.CharType;
@@ -545,7 +544,6 @@ public class OrcTester
                         new int[] {0},
                         new int[] {0},
                         true,
-                        new AbstractFilterFunction[0],
                         512 * 1024);
                 recordReader.pushdownFilterAndProjection(options, new int[] {0}, ImmutableList.of(type), new Block[1]);
